refactor(uex): extract snapshot reconciliation from PriceSyncer
Applying a UEX snapshot and working out which of our leftovers are genuinely
gone rather than merely missing from a short answer is the same problem
whether the rows are ships or cargo. The retention guard in particular is
subtle enough that a second copy would be a second thing to get wrong.

Behaviour is unchanged -- PriceSyncer's 26 tests pass against it.
